accelerometer_value_correction
trying to correct the errors in the accelerometer readings, due vibrations over a short period of time, using a "fixed-value" matrix. this will probably not work for all accel. values, and a filter (Kalman) might be implemented to account for variations in the errors. accel. readings suffer short-term vibrations that cause readings over a differential to be off, as opposed to gyrsocopes that are usually accurate over short durations but eventually drift with time.