change interrupt detection
due to the nature of the project, some form of interrupt (hardware/software) needs to be triggered whenever there is a "significant" change in the input from the receiver (paired to an RF controller).

in this case, i am using the Arduino Uno's PinChangeInterrupt Vectors; this groups certain pins together and accounts for when any of the member pin's values change. an interrupt can also be set up, individually, for each pin within an interrupt vector using the corresponding bit index of each pin in either a PINB or a PINC port.

whenever a change is detected for a PCI_vect, an interrupt is triggered and a subroutine is called for that PCI_vect, where the state of its member pins can be checked against the bits in a PINB or PINC register as mentioned above. the duration for which the values changed can be measured using the Arduino micros() function, as a timestamp.

more info at: http://www.atmel.com/images/Atmel-8271-8-bit-AVR-Microcontroller-ATmega48A-48PA-88A-88PA-168A-168